refactor: profile 폴더 정리 및 recoil로 유저세션관리
