+++ src/api/Chat.api.ts
@@ -2,8 +2,8 @@ import { useMutation, useQueryClient } from 'react-query';
 import { supabase } from './Supabase.api';
 
 import { useRecoilState } from 'recoil';
-import { ChatId } from '../atom/Product.atom';
 import { ChatDataType } from '../components/chat/ChatModal.type';
+import { ChatId } from '../state/atom/chatAtom';
 import { CreateOrGetChatTypes, SendDirectMessageTypes } from './Api.type';
 
 // 챗방 생성 또는 가져오기 로직을 커스텀 훅으로 분리
