+++ src/components/profile/UserProfile.tsx
@@ -1,24 +1,24 @@
-import { Session } from '@supabase/supabase-js';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { getUserDataHandler, getUserSessionHandler } from '../../api/Supabase.api';
+import { useRecoilValue } from 'recoil';
+import { getUserDataHandler } from '../../api/Supabase.api';
 import profileImage from '../../assets/profile/defaultprofileimage.webp';
+import { userSession } from '../../state/atom/userSessionAtom';
 import { Tables } from '../../types/types';
 import Follow from '../common/follow/Follow';
-import Tab from './Tab';
 import * as St from './UserProfile.styled';
+import Tab from './tab/Tab';
 
 const UserProfile = () => {
   const params = useParams().id;
-  const [userSession, setUserSession] = useState<Session | null>(null);
   const [userData, setUserData] = useState<Tables<'users'>>();
   const [nickname, setNickname] = useState('');
+  const session = useRecoilValue(userSession);
+
   const getUserData = async () => {
     const result = await getUserDataHandler(params as string);
-    const session = await getUserSessionHandler();
     setUserData(result[0]);
     setNickname(result[0].nickname);
-    setUserSession(session.session);
   };
 
   useEffect(() => {
@@ -30,16 +30,13 @@ const UserProfile = () => {
       <St.Title>프로필</St.Title>
       <St.ProfileWrapper>
         <St.ProfileImg src={userData?.user_img || profileImage} />
-        {userSession?.user.id === params ? (
+        {session?.id === params ? (
           // 마이 프로필
           <St.ProfileInfo>
             <St.ProfileNicknameEmail>
               <St.ProfileNickname>
                 안녕하세요!{' '}
-                <St.ProfileUserName>
-                  {nickname || userSession?.user.user_metadata.preferred_username}
-                </St.ProfileUserName>
-                님
+                <St.ProfileUserName>{nickname || session?.user_metadata.preferred_username}</St.ProfileUserName>님
               </St.ProfileNickname>
               <St.ProfileIntroText>{userData?.intro_text}</St.ProfileIntroText>
             </St.ProfileNicknameEmail>
@@ -53,7 +50,7 @@ const UserProfile = () => {
           </St.TargetProfileInfo>
         )}
       </St.ProfileWrapper>
-      <Tab userSession={userSession} userData={userData} />
+      <Tab />
     </St.Container>
   );
 };
